Added the #once method, and options for #on
It is now possible to give an object of options to the #on method. The only recognized parameter implemented yet is the "once", which makes the bound function being unbound after the first time it is executed.
Also added a shortcut, the #once method, that will call the #on method with an object of options with only the "once" flag set to true.
Finally, some very basic unit tests about all that have been added.